% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.Helicini.R
\name{export.Helicini}
\alias{export.Helicini}
\title{Export Helicini dataset into your working directory.}
\usage{
export.Helicini()
}
\value{
Creates an directory './Helicini_data' and copies the full mitochondrial Helicini dataset with metadata into this directory.
}
\description{
Exports the complete mitochondrial Helicini dataset into your working directory.
}
