% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative.alignment.R
\name{representative.alignment}
\alias{representative.alignment}
\title{Export representative mitochondrial alignment for the Helicini tribe allowing for choosing divergence level, loci, and individuals.}
\usage{
representative.alignment(
  level = c("ID", "subclade", "clade", "genus"),
  genes = c("COX1", "TRNV", "16S", "TRNL1_TRNA", "ND6", "TRNP", "ND5", "ND1_ND4L",
    "CYTB", "TRND-TRNF", "COX2", "TRNY-TRNL2", "ATP8", "TRNN", "ATP6", "TRNR_TRNE",
    "12S_TRNM", "ND3", "TRNS2_TRNT", "COX3", "TRNS1", "ND4", "TRNI", "ND2", "TRNK"),
  genus = NULL,
  species = NULL,
  individuals = NULL,
  weight = c("max", "average"),
  maxloci = FALSE,
  save = TRUE,
  file = "representative.fas",
  save_single = FALSE
)
}
\arguments{
\item{level}{character. Divergence level to be represented. See details for allowed values. Only a single level is accepted. Using ID would export all individuals (see also "individuals").}

\item{genes}{character. Loci to be included. See details for allowed values.}

\item{genus}{character. Genera to be considered. If not present in the data, they will be ignored without warning.}

\item{species}{character. Species to be considered. If not present in the data, they will be ignored without warning.}

\item{individuals}{character. Optional, vector listing individual IDs to be considered in the selection. Can also be a data.frame, but then it is assumed that the IDs are in the first column.}

\item{weight}{character. Either "max" or "average" (use "max" as default). Specifies the scoring mechanism used, see details.}

\item{maxloci}{Boolean. If TRUE, the individual with best coverage (see details) is selected as usual, but then additional individuals are added if these help to increase the number of loci available for each of the groups. Default FALSE.}

\item{save}{Boolean. Save the final alignment to file? Default TRUE.}

\item{file}{character. Name for the file containing the exported representative alignment.}

\item{save_single}{Boolean. Specifies whether to save single gene alignments ("best.alignments" in the function output. Default FALSE.).}
}
\value{
Returns a list of the following objects: "per.group.inds","lengths.ordered.groups","missing.loci","best.alignments","final.alignment". "per.group.inds" gives the number of individuals in each group specified by "level" parameter. 'missing.loci' has three values: 0=locus present in the best scoring individual, 1=locus missing in the best scoring individual ,2=locus missing for the group altogether. "lengths.ordered.groups" contains all the sequence lengths structured by group and locus.
}
\description{
Finds the individual with most complete data for each clade at the specified divergence level, concatenates the specified loci and outputs the alignment. Useful when you want to show the phylogenetic position of your new sample relative to what is already known, but want to reduce the data. In output it provides some useful info about the data.
}
\details{
You may choose from a list of alignment files to output only selected loci. By default, all are exported in the order as they appear in the mitogenomes ("COX1","TRNV","16S","TRNL1_TRNA","ND6","TRNP","ND5","ND1_ND4L","CYTB","TRND-TRNF","COX2","TRNY-TRNL2","ATP8","TRNN","ATP6","TRNR_TRNE","12S_TRNM","ND3","TRNS2_TRNT","COX3","TRNS1","ND4","TRNI","ND2","TRNK"). Running with "loci" at default value outputs complete mitogenomes where available (with gaps and question marks for missing data in part of or the whole locus, respectively).

"level" has three levels, corresponding to Table S1 in Korábek et al. 2022: 'genus' for genus, 'clade' for a roughly species-level clades (these clades do not always correspond to species borders), and 'subclade' for intraspecific clades used in figs. S4-62.

The individuals are selected as follows: Samples within each clade specified by "level" are ranked for the sequence length of each locus. Loci are then weighted by their alignment length or by average non-zero sequence length and the individual with the greatest weighted sequence length is chosen as representative of its clade.

The "maxloci" option, if on, outputs the individual with best coverage overall, but if some locus is not available for this individual, it looks for an individual that has it (the one with the longest sequence) and adds also this individual to the final alignment.
}
