% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{selector}
\alias{selector}
\title{Select sequences from an alignment}
\usage{
selector(individuals, alignment, save = T, filename)
}
\arguments{
\item{individuals}{character. List of sequence identifiers to be selected}

\item{alignment}{character vector or data.frame. Source alignment in sequential fasta format.}

\item{save}{Boolean. Save the selected sequences in a file? Default TRUE.}

\item{filename}{character. Name of the file to be saved.}
}
\description{
Selects sequences from alignment using a vector of sequence identifiers.
}
\examples{
\dontrun{
 tab<-read_ods("210917_tabulka_mol_kopie.ods",col_names = TRUE,col_types = NA,sheet="Helicini")
 individuals<-tab$ID
}
}
