% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs_lengths.R
\name{seqs_lengths}
\alias{seqs_lengths}
\title{Calculate and plot the distribution of sequence lengths in an alignment}
\usage{
seqs_lengths(alignment, ambiguous = FALSE, verbose = TRUE)
}
\arguments{
\item{alignment}{vector or data.frame. Alignment in sequential fasta format.}

\item{ambiguous}{Boolean. Specifies if ambiguous bases should be included in the sequence length, default FALSE.}

\item{verbose}{Boolean. If TRUE, length statistics are printed to the screen and the length distribution plotted; default TRUE.}
}
\value{
Produces a named vector of sequence lengths.
}
\description{
Extracts the lengths of sequences in an alignment, returns a named vector of the lengths and plots their distribution
}
