% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.iso.R
\name{conc.iso}
\alias{conc.iso}
\title{Concatenate sequences based on identifiers in the fasta header.}
\usage{
conc.iso(alignments, save = T, file = "concatenate.fas", uknown = "?")
}
\arguments{
\item{alignments}{list. List containing fasta sequence alignments loaded in as data.frame or character vector by read.table.}

\item{save}{Boolean. If TRUE, the results will be saved into a fasta file.}

\item{file}{character. Name of the file containing the concatenated alignment, if save=TRUE)}

\item{uknown}{character. The character used to denote missing data when the individual is not represented by a given locus. Does affect coding of missing data in the concatenated alignments. Default "?", other meaningful options are "-" (same as gap) or "N" (it is assumed that there is a nucleotide).}
}
\value{
Returns concatenated alignment in a sequential FASTA format as an character vector.
}
\description{
Performs concatenation of sequences using 8-letter identifiers at the beginning of fasta headers.
}
